requirements.txt
Babel==2.2.0
Django==1.9.2
django-multiselectfield==0.1.3
django-phonenumber-field==1.0.0
django-tastypie==0.13.3
phonenumbers==7.2.6
python-dateutil==2.4.2
python-mimeparse==1.5.1
pytz==2015.7
six==1.10.0
wheel==0.24.0
